feat(server): add status field to task creation with default value feat(client): improve styling and layout of ProjectInfo component
