+++ apps/client/src/components/KanbanBoard.tsx
@@ -3,22 +3,11 @@ import { DndProvider, useDrag, useDrop } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 import Projectinfo from "./ProjectInfo";
 import axios from "axios";
-
 import { useAuth } from "@clerk/clerk-react";
+import AddTaskModal, { TaskStatus, Priority } from "./AddTaskModal"; // Import the AddTaskModal component
 
 // Define enums and interfaces based on the schema
-enum TaskStatus {
-  TODO = "TODO",
-  IN_PROGRESS = "IN_PROGRESS",
-  DONE = "DONE",
-  ARCHIVED = "ARCHIVED",
-}
-
-enum Priority {
-  HIGH = "HIGH",
-  MEDIUM = "MEDIUM",
-  LOW = "LOW",
-}
+// Note: TaskStatus and Priority are now imported from AddTaskModal
 
 interface User {
   id: string;
@@ -82,7 +71,7 @@ const KanbanColumn: React.FC<ColumnProps> = ({
     <div
       ref={ref}
       className={`flex-1 min-w-[300px] h-full ${
-        darkMode ? "bg-[#1E1E1E]" : "bg-white"
+        darkMode ? "bg-[#171717]" : "bg-white"
       }`}
     >
       <div
@@ -216,261 +205,6 @@ const TaskCard: React.FC<TaskCardProps> = ({ task, darkMode }) => {
 interface KanbanBoardProps {
   projectId: string | null;
 }
-
-// Add Task Modal Component
-interface AddTaskModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  darkMode: boolean;
-  projectId: string;
-  onTaskAdded: () => void; // Add this new callback prop
-}
-
-const AddTaskModal: React.FC<AddTaskModalProps> = ({
-  isOpen,
-  onClose,
-  darkMode,
-  projectId,
-  onTaskAdded, // Add this new prop
-}) => {
-  const [formData, setFormData] = useState({
-    title: "",
-    description: "",
-    priority: Priority.MEDIUM,
-    dueDate: "",
-    type: "TASK",
-  });
-  const [isLoading, setIsLoading] = useState(false); // Add loading state
-  const { getToken } = useAuth();
-
-  //create task endpoint
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true); // Start loading
-
-    try {
-      const token = await getToken();
-      console.log(
-        "Token obtained:",
-        token ? `${token.substring(0, 15)}...` : "No token"
-      );
-
-      console.log("Sending request to create task with data:", {
-        ...formData,
-        projectId,
-      });
-
-      const response = await axios.post(
-        "http://localhost:5000/api/tasks/create",
-        {
-          ...formData,
-          projectId: projectId,
-        },
-        {
-          headers: { Authorization: `Bearer ${token}` },
-        }
-      );
-
-      if (response.status === 201) {
-        console.log("Task created successfully:", response.data);
-        // Clear form data to reset the form
-        setFormData({
-          title: "",
-          description: "",
-          priority: Priority.MEDIUM,
-          dueDate: "",
-          type: "TASK",
-        });
-        // Call the callback to refresh tasks
-        onTaskAdded();
-        onClose();
-      }
-    } catch (error) {
-      console.error("Failed to create task - Full error details:", error);
-
-      if (axios.isAxiosError(error)) {
-        console.error("Request failed with status:", error.response?.status);
-        console.error("Error message:", error.message);
-        console.error("Server response data:", error.response?.data);
-        console.error("Request URL:", error.config?.url);
-        console.error("Request method:", error.config?.method);
-        console.error("Request headers:", error.config?.headers);
-      }
-    } finally {
-      setIsLoading(false); // End loading whether request succeeded or failed
-    }
-  };
-
-  if (!isOpen) return null;
-
-  const inputStyles = `w-full p-2 rounded border focus:outline-none focus:ring-2 focus:ring-emerald-400 ${
-    darkMode
-      ? "bg-[#1E1E1E] border-[#2C2C2C] text-white"
-      : "bg-white border-gray-200 text-[#212121]"
-  }`;
-
-  return (
-    <div className="fixed inset-0 flex items-center justify-center z-50">
-      <div
-        className="absolute inset-0 backdrop-blur-md bg-black/30"
-        onClick={onClose}
-      ></div>
-      <div
-        className={`relative w-full max-w-md transform transition-all duration-300 ease-in-out ${
-          darkMode
-            ? "bg-[#171717]/95 border border-[#2C2C2C]"
-            : "bg-gray-100/95"
-        } rounded-lg shadow-xl backdrop-blur-sm`}
-      >
-        <div className="p-8">
-          <div className="flex justify-between items-center mb-6">
-            <h2
-              className={`text-2xl font-bold ${
-                darkMode ? "text-white" : "text-[#212121]"
-              }`}
-            >
-              Add New Task
-            </h2>
-            <button
-              onClick={onClose}
-              className={`p-2 rounded-full hover:bg-opacity-80 ${
-                darkMode
-                  ? "text-gray-400 hover:bg-[#2C2C2C]"
-                  : "text-gray-600 hover:bg-gray-200"
-              }`}
-            >
-              <svg
-                className="w-5 h-5"
-                fill="none"
-                stroke="currentColor"
-                viewBox="0 0 24 24"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth="2"
-                  d="M6 18L18 6M6 6l12 12"
-                />
-              </svg>
-            </button>
-          </div>
-          <form onSubmit={handleSubmit} className="space-y-6">
-            <div>
-              <label
-                className={`block mb-2 font-medium ${
-                  darkMode ? "text-gray-300" : "text-gray-700"
-                }`}
-              >
-                Title <span className="text-emerald-400">*</span>
-              </label>
-              <input
-                type="text"
-                required
-                value={formData.title}
-                onChange={(e) =>
-                  setFormData({ ...formData, title: e.target.value })
-                }
-                className={inputStyles}
-                placeholder="Enter task title"
-              />
-            </div>
-            <div>
-              <label
-                className={`block mb-2 font-medium ${
-                  darkMode ? "text-gray-300" : "text-gray-700"
-                }`}
-              >
-                Description
-              </label>
-              <textarea
-                value={formData.description}
-                onChange={(e) =>
-                  setFormData({ ...formData, description: e.target.value })
-                }
-                className={inputStyles}
-                rows={3}
-                placeholder="Enter task description"
-              />
-            </div>
-            <div>
-              <label
-                className={`block mb-2 font-medium ${
-                  darkMode ? "text-gray-300" : "text-gray-700"
-                }`}
-              >
-                Priority
-              </label>
-              <select
-                value={formData.priority}
-                onChange={(e) =>
-                  setFormData({
-                    ...formData,
-                    priority: e.target.value as Priority,
-                  })
-                }
-                className={inputStyles}
-              >
-                <option value={Priority.LOW}>Low</option>
-                <option value={Priority.MEDIUM}>Medium</option>
-                <option value={Priority.HIGH}>High</option>
-              </select>
-            </div>
-            <div>
-              <label
-                className={`block mb-2 font-medium ${
-                  darkMode ? "text-gray-300" : "text-gray-700"
-                }`}
-              >
-                Due Date
-              </label>
-              <input
-                type="date"
-                value={formData.dueDate}
-                onChange={(e) =>
-                  setFormData({ ...formData, dueDate: e.target.value })
-                }
-                className={inputStyles}
-              />
-            </div>
-            <button
-              type="submit"
-              disabled={isLoading}
-              className="w-full py-3 px-4 rounded-lg bg-emerald-400 hover:bg-emerald-500 text-white font-medium transition-colors duration-200 flex items-center justify-center"
-            >
-              {isLoading ? (
-                <>
-                  <svg
-                    className="animate-spin -ml-1 mr-3 h-5 w-5 text-white"
-                    xmlns="http://www.w3.org/2000/svg"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                  >
-                    <circle
-                      className="opacity-25"
-                      cx="12"
-                      cy="12"
-                      r="10"
-                      stroke="currentColor"
-                      strokeWidth="4"
-                    ></circle>
-                    <path
-                      className="opacity-75"
-                      fill="currentColor"
-                      d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-                    ></path>
-                  </svg>
-                  Creating...
-                </>
-              ) : (
-                "Create Task"
-              )}
-            </button>
-          </form>
-        </div>
-      </div>
-    </div>
-  );
-};
 
 // Main Kanban Board Component
 const KanbanBoard: React.FC<KanbanBoardProps> = ({ projectId }) => {
