+++ apps/server/src/controllers/taskController.js
@@ -1,10 +1,9 @@
-
 import prisma from '../db.js'
 
 // Create a new task
 export const createTask = async (req, res) => {
   try {
-    const { title, description, projectId, type, priority, dueDate } = req.body;
+    const { title, description, projectId, type, priority, dueDate, status } = req.body;
     const {emailAddresses}= req.auth 
     const email = emailAddresses?.[0]?.emailAddress;
 
@@ -48,6 +47,7 @@ export const createTask = async (req, res) => {
         projectId,
         type: type || 'TASK', // Default to TASK if not provided
         priority: priority || 'MEDIUM', // Default to MEDIUM
+        status: status || 'TODO', // Default to TODO if not provided
         dueDate: dueDate ? new Date(dueDate) : null,
         creatorId: user.id, // Set the creator
       },
